requirements.txt
Pillow>=10.0
pycdlib>=1.14
numpy>=1.24
